[FLINK-6887] [table] Split up CodeGenerator into several specific CodeGenerator
This closes #4171.
