+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/codegen/CodeGenerator.scala
@@ -18,8 +18,6 @@
 
 package org.apache.flink.table.codegen
 
-import java.lang.reflect.ParameterizedType
-import java.lang.{Iterable => JIterable}
 import java.math.{BigDecimal => JBigDecimal}
 
 import org.apache.calcite.avatica.util.DateTimeUtils
@@ -28,25 +26,21 @@ import org.apache.calcite.sql.SqlOperator
 import org.apache.calcite.sql.`type`.SqlTypeName._
 import org.apache.calcite.sql.fun.SqlStdOperatorTable._
 import org.apache.flink.api.common.functions._
-import org.apache.flink.api.common.io.GenericInputFormat
 import org.apache.flink.api.common.typeinfo._
 import org.apache.flink.api.common.typeutils.CompositeType
 import org.apache.flink.api.java.typeutils._
 import org.apache.flink.api.scala.typeutils.CaseClassTypeInfo
-import org.apache.flink.configuration.Configuration
 import org.apache.flink.streaming.api.functions.ProcessFunction
 import org.apache.flink.table.api.TableConfig
 import org.apache.flink.table.calcite.FlinkTypeFactory
 import org.apache.flink.table.codegen.CodeGenUtils._
 import org.apache.flink.table.codegen.GeneratedExpression.{NEVER_NULL, NO_CODE}
-import org.apache.flink.table.codegen.Indenter.toISC
 import org.apache.flink.table.codegen.calls.{BuiltInMethods, FunctionGenerator}
 import org.apache.flink.table.codegen.calls.ScalarOperators._
 import org.apache.flink.table.functions.sql.ScalarSqlFunctions
 import org.apache.flink.table.functions.utils.UserDefinedFunctionUtils
 import org.apache.flink.table.functions.utils.UserDefinedFunctionUtils.{getUserDefinedMethod, signatureToString}
-import org.apache.flink.table.functions.{AggregateFunction, FunctionContext, TimeMaterializationSqlFunction, UserDefinedFunction}
-import org.apache.flink.table.runtime.TableFunctionCollector
+import org.apache.flink.table.functions.{FunctionContext, TimeMaterializationSqlFunction, UserDefinedFunction}
 import org.apache.flink.table.typeutils.TypeCheckUtils._
 import org.apache.flink.types.Row
 
@@ -54,7 +48,9 @@ import scala.collection.JavaConversions._
 import scala.collection.mutable
 
 /**
-  * A code generator for generating Flink [[org.apache.flink.api.common.functions.Function]]s.
+  * [[CodeGenerator]] is the base code generator for generating Flink
+  * [[org.apache.flink.api.common.functions.Function]]s.
+  * It is responsible for expression generation and tracks the context (member variables etc).
   *
   * @param config configuration that determines runtime behavior
   * @param nullableInput input(s) can be null.
@@ -65,7 +61,7 @@ import scala.collection.mutable
   * @param input2FieldMapping additional mapping information for input2
   *   (e.g. POJO types have no deterministic field order and some input fields might not be read)
   */
-class CodeGenerator(
+abstract class CodeGenerator(
     config: TableConfig,
     nullableInput: Boolean,
     input1: TypeInformation[_ <: Any],
@@ -95,43 +91,18 @@ class CodeGenerator(
     case _ => // ok
   }
 
-  private val input1Mapping = input1FieldMapping match {
+  protected val input1Mapping: Array[Int] = input1FieldMapping match {
     case Some(mapping) => mapping
     case _ => (0 until input1.getArity).toArray
   }
 
-  private val input2Mapping = input2FieldMapping match {
+  protected val input2Mapping: Array[Int] = input2FieldMapping match {
     case Some(mapping) => mapping
     case _ => input2 match {
       case Some(input) => (0 until input.getArity).toArray
       case _ => Array[Int]()
     }
   }
-
-  /**
-    * A code generator for generating unary Flink
-    * [[org.apache.flink.api.common.functions.Function]]s with one input.
-    *
-    * @param config configuration that determines runtime behavior
-    * @param nullableInput input(s) can be null.
-    * @param input type information about the input of the Function
-    * @param inputFieldMapping additional mapping information necessary for input
-    *   (e.g. POJO types have no deterministic field order and some input fields might not be read)
-    */
-  def this(
-      config: TableConfig,
-      nullableInput: Boolean,
-      input: TypeInformation[Any],
-      inputFieldMapping: Array[Int]) =
-    this(config, nullableInput, input, None, Some(inputFieldMapping))
-
-  /**
-    * A code generator for generating Flink input formats.
-    *
-    * @param config configuration that determines runtime behavior
-    */
-  def this(config: TableConfig) =
-    this(config, false, new RowTypeInfo(), None, None)
 
   // set of member statements that will be added only once
   // we use a LinkedHashSet to keep the insertion order
@@ -258,604 +229,6 @@ class CodeGenerator(
     */
   def generateExpression(rex: RexNode): GeneratedExpression = {
     rex.accept(this)
-  }
-
-  /**
-    * Generates a [[org.apache.flink.table.runtime.aggregate.GeneratedAggregations]] that can be
-    * passed to a Java compiler.
-    *
-    * @param name        Class name of the function.
-    *                    Does not need to be unique but has to be a valid Java class identifier.
-    * @param generator   The code generator instance
-    * @param physicalInputTypes Physical input row types
-    * @param aggregates  All aggregate functions
-    * @param aggFields   Indexes of the input fields for all aggregate functions
-    * @param aggMapping  The mapping of aggregates to output fields
-    * @param partialResults A flag defining whether final or partial results (accumulators) are set
-    *                       to the output row.
-    * @param fwdMapping  The mapping of input fields to output fields
-    * @param mergeMapping An optional mapping to specify the accumulators to merge. If not set, we
-    *                     assume that both rows have the accumulators at the same position.
-    * @param constantFlags An optional parameter to define where to set constant boolean flags in
-    *                      the output row.
-    * @param outputArity The number of fields in the output row.
-    * @param needRetract a flag to indicate if the aggregate needs the retract method
-    * @param needMerge a flag to indicate if the aggregate needs the merge method
-    * @param needReset a flag to indicate if the aggregate needs the resetAccumulator method
-    *
-    * @return A GeneratedAggregationsFunction
-    */
-  def generateAggregations(
-      name: String,
-      generator: CodeGenerator,
-      physicalInputTypes: Seq[TypeInformation[_]],
-      aggregates: Array[AggregateFunction[_ <: Any, _ <: Any]],
-      aggFields: Array[Array[Int]],
-      aggMapping: Array[Int],
-      partialResults: Boolean,
-      fwdMapping: Array[Int],
-      mergeMapping: Option[Array[Int]],
-      constantFlags: Option[Array[(Int, Boolean)]],
-      outputArity: Int,
-      needRetract: Boolean,
-      needMerge: Boolean,
-      needReset: Boolean)
-  : GeneratedAggregationsFunction = {
-
-    // get unique function name
-    val funcName = newName(name)
-    // register UDAGGs
-    val aggs = aggregates.map(a => generator.addReusableFunction(a))
-    // get java types of accumulators
-    val accTypeClasses = aggregates.map { a =>
-      a.getClass.getMethod("createAccumulator").getReturnType
-    }
-    val accTypes = accTypeClasses.map(_.getCanonicalName)
-
-    // get java classes of input fields
-    val javaClasses = physicalInputTypes.map(t => t.getTypeClass)
-    // get parameter lists for aggregation functions
-    val parameters = aggFields.map { inFields =>
-      val fields = for (f <- inFields) yield
-        s"(${javaClasses(f).getCanonicalName}) input.getField($f)"
-      fields.mkString(", ")
-    }
-    val methodSignaturesList = aggFields.map {
-      inFields => for (f <- inFields) yield javaClasses(f)
-    }
-
-    // check and validate the needed methods
-    aggregates.zipWithIndex.map {
-      case (a, i) => {
-        getUserDefinedMethod(a, "accumulate", Array(accTypeClasses(i)) ++ methodSignaturesList(i))
-          .getOrElse(
-            throw new CodeGenException(
-              s"No matching accumulate method found for AggregateFunction " +
-                s"'${a.getClass.getCanonicalName}'" +
-                s"with parameters '${signatureToString(methodSignaturesList(i))}'.")
-          )
-
-        if (needRetract) {
-          getUserDefinedMethod(a, "retract", Array(accTypeClasses(i)) ++ methodSignaturesList(i))
-            .getOrElse(
-              throw new CodeGenException(
-                s"No matching retract method found for AggregateFunction " +
-                  s"'${a.getClass.getCanonicalName}'" +
-                  s"with parameters '${signatureToString(methodSignaturesList(i))}'.")
-            )
-        }
-
-        if (needMerge) {
-          val methods =
-            getUserDefinedMethod(a, "merge", Array(accTypeClasses(i), classOf[JIterable[Any]]))
-              .getOrElse(
-                throw new CodeGenException(
-                  s"No matching merge method found for AggregateFunction " +
-                    s"${a.getClass.getCanonicalName}'.")
-              )
-
-          var iterableTypeClass = methods.getGenericParameterTypes.apply(1)
-            .asInstanceOf[ParameterizedType].getActualTypeArguments.apply(0)
-          // further extract iterableTypeClass if the accumulator has generic type
-          iterableTypeClass match {
-            case impl: ParameterizedType => iterableTypeClass = impl.getRawType
-            case _ =>
-          }
-
-          if (iterableTypeClass != accTypeClasses(i)) {
-            throw new CodeGenException(
-              s"merge method in AggregateFunction ${a.getClass.getCanonicalName} does not have " +
-                s"the correct Iterable type. Actually: ${iterableTypeClass.toString}. " +
-                s"Expected: ${accTypeClasses(i).toString}")
-          }
-        }
-
-        if (needReset) {
-          getUserDefinedMethod(a, "resetAccumulator", Array(accTypeClasses(i)))
-            .getOrElse(
-              throw new CodeGenException(
-                s"No matching resetAccumulator method found for " +
-                  s"aggregate ${a.getClass.getCanonicalName}'.")
-            )
-        }
-      }
-    }
-
-    def genSetAggregationResults: String = {
-
-      val sig: String =
-        j"""
-           |  public final void setAggregationResults(
-           |    org.apache.flink.types.Row accs,
-           |    org.apache.flink.types.Row output)""".stripMargin
-
-      val setAggs: String = {
-        for (i <- aggs.indices) yield
-
-          if (partialResults) {
-            j"""
-               |    output.setField(
-               |      ${aggMapping(i)},
-               |      (${accTypes(i)}) accs.getField($i));""".stripMargin
-          } else {
-            j"""
-               |    org.apache.flink.table.functions.AggregateFunction baseClass$i =
-               |      (org.apache.flink.table.functions.AggregateFunction) ${aggs(i)};
-               |
-               |    output.setField(
-               |      ${aggMapping(i)},
-               |      baseClass$i.getValue((${accTypes(i)}) accs.getField($i)));""".stripMargin
-          }
-      }.mkString("\n")
-
-      j"""
-         |$sig {
-         |$setAggs
-         |  }""".stripMargin
-    }
-
-    def genAccumulate: String = {
-
-      val sig: String =
-        j"""
-            |  public final void accumulate(
-            |    org.apache.flink.types.Row accs,
-            |    org.apache.flink.types.Row input)""".stripMargin
-
-      val accumulate: String = {
-        for (i <- aggs.indices) yield
-          j"""
-             |    ${aggs(i)}.accumulate(
-             |      ((${accTypes(i)}) accs.getField($i)),
-             |      ${parameters(i)});""".stripMargin
-      }.mkString("\n")
-
-      j"""$sig {
-         |$accumulate
-         |  }""".stripMargin
-    }
-
-    def genRetract: String = {
-
-      val sig: String =
-        j"""
-            |  public final void retract(
-            |    org.apache.flink.types.Row accs,
-            |    org.apache.flink.types.Row input)""".stripMargin
-
-      val retract: String = {
-        for (i <- aggs.indices) yield
-          j"""
-             |    ${aggs(i)}.retract(
-             |      ((${accTypes(i)}) accs.getField($i)),
-             |      ${parameters(i)});""".stripMargin
-      }.mkString("\n")
-
-      if (needRetract) {
-        j"""
-           |$sig {
-           |$retract
-           |  }""".stripMargin
-      } else {
-        j"""
-           |$sig {
-           |  }""".stripMargin
-      }
-    }
-
-    def genCreateAccumulators: String = {
-
-      val sig: String =
-        j"""
-           |  public final org.apache.flink.types.Row createAccumulators()
-           |    """.stripMargin
-      val init: String =
-        j"""
-           |      org.apache.flink.types.Row accs =
-           |          new org.apache.flink.types.Row(${aggs.length});"""
-          .stripMargin
-      val create: String = {
-        for (i <- aggs.indices) yield
-          j"""
-             |    accs.setField(
-             |      $i,
-             |      ${aggs(i)}.createAccumulator());"""
-            .stripMargin
-      }.mkString("\n")
-      val ret: String =
-        j"""
-           |      return accs;"""
-          .stripMargin
-
-      j"""$sig {
-         |$init
-         |$create
-         |$ret
-         |  }""".stripMargin
-    }
-
-    def genSetForwardedFields: String = {
-
-      val sig: String =
-        j"""
-           |  public final void setForwardedFields(
-           |    org.apache.flink.types.Row input,
-           |    org.apache.flink.types.Row output)
-           |    """.stripMargin
-
-      val forward: String = {
-        for (i <- fwdMapping.indices if fwdMapping(i) >= 0) yield
-          {
-            j"""
-               |    output.setField(
-               |      $i,
-               |      input.getField(${fwdMapping(i)}));"""
-              .stripMargin
-          }
-      }.mkString("\n")
-
-      j"""$sig {
-         |$forward
-         |  }""".stripMargin
-    }
-
-    def genSetConstantFlags: String = {
-
-      val sig: String =
-        j"""
-           |  public final void setConstantFlags(org.apache.flink.types.Row output)
-           |    """.stripMargin
-
-      val setFlags: String = if (constantFlags.isDefined) {
-        {
-          for (cf <- constantFlags.get) yield {
-            j"""
-               |    output.setField(${cf._1}, ${if (cf._2) "true" else "false"});"""
-              .stripMargin
-          }
-        }.mkString("\n")
-      } else {
-        ""
-      }
-
-      j"""$sig {
-         |$setFlags
-         |  }""".stripMargin
-    }
-
-    def genCreateOutputRow: String = {
-      j"""
-         |  public final org.apache.flink.types.Row createOutputRow() {
-         |    return new org.apache.flink.types.Row($outputArity);
-         |  }""".stripMargin
-    }
-
-    def genMergeAccumulatorsPair: String = {
-
-      val mapping = mergeMapping.getOrElse(aggs.indices.toArray)
-
-      val sig: String =
-        j"""
-           |  public final org.apache.flink.types.Row mergeAccumulatorsPair(
-           |    org.apache.flink.types.Row a,
-           |    org.apache.flink.types.Row b)
-           """.stripMargin
-      val merge: String = {
-        for (i <- aggs.indices) yield
-          j"""
-             |    ${accTypes(i)} aAcc$i = (${accTypes(i)}) a.getField($i);
-             |    ${accTypes(i)} bAcc$i = (${accTypes(i)}) b.getField(${mapping(i)});
-             |    accIt$i.setElement(bAcc$i);
-             |    ${aggs(i)}.merge(aAcc$i, accIt$i);
-             |    a.setField($i, aAcc$i);
-             """.stripMargin
-      }.mkString("\n")
-      val ret: String =
-        j"""
-           |      return a;
-           """.stripMargin
-
-      if (needMerge) {
-        j"""
-           |$sig {
-           |$merge
-           |$ret
-           |  }""".stripMargin
-      } else {
-        j"""
-           |$sig {
-           |$ret
-           |  }""".stripMargin
-      }
-    }
-
-    def genMergeList: String = {
-      {
-        val singleIterableClass = "org.apache.flink.table.runtime.aggregate.SingleElementIterable"
-        for (i <- accTypes.indices) yield
-          j"""
-             |    private final $singleIterableClass<${accTypes(i)}> accIt$i =
-             |      new $singleIterableClass<${accTypes(i)}>();
-             """.stripMargin
-      }.mkString("\n")
-    }
-
-    def genResetAccumulator: String = {
-
-      val sig: String =
-        j"""
-           |  public final void resetAccumulator(
-           |    org.apache.flink.types.Row accs)""".stripMargin
-
-      val reset: String = {
-        for (i <- aggs.indices) yield
-          j"""
-             |    ${aggs(i)}.resetAccumulator(
-             |      ((${accTypes(i)}) accs.getField($i)));""".stripMargin
-      }.mkString("\n")
-
-      if (needReset) {
-        j"""$sig {
-           |$reset
-           |  }""".stripMargin
-      } else {
-        j"""$sig {
-           |  }""".stripMargin
-      }
-    }
-
-    var funcCode =
-      j"""
-         |public final class $funcName
-         |  extends org.apache.flink.table.runtime.aggregate.GeneratedAggregations {
-         |
-         |  ${reuseMemberCode()}
-         |  $genMergeList
-         |  public $funcName() throws Exception {
-         |    ${reuseInitCode()}
-         |  }
-         |  ${reuseConstructorCode(funcName)}
-         |
-         """.stripMargin
-
-    funcCode += genSetAggregationResults + "\n"
-    funcCode += genAccumulate + "\n"
-    funcCode += genRetract + "\n"
-    funcCode += genCreateAccumulators + "\n"
-    funcCode += genSetForwardedFields + "\n"
-    funcCode += genSetConstantFlags + "\n"
-    funcCode += genCreateOutputRow + "\n"
-    funcCode += genMergeAccumulatorsPair + "\n"
-    funcCode += genResetAccumulator + "\n"
-    funcCode += "}"
-
-    GeneratedAggregationsFunction(funcName, funcCode)
-  }
-
-  /**
-    * Generates a [[org.apache.flink.api.common.functions.Function]] that can be passed to Java
-    * compiler.
-    *
-    * @param name Class name of the Function. Must not be unique but has to be a valid Java class
-    *             identifier.
-    * @param clazz Flink Function to be generated.
-    * @param bodyCode code contents of the SAM (Single Abstract Method). Inputs, collector, or
-    *                 output record can be accessed via the given term methods.
-    * @param returnType expected return type
-    * @tparam F Flink Function to be generated.
-    * @tparam T Return type of the Flink Function.
-    * @return instance of GeneratedFunction
-    */
-  def generateFunction[F <: Function, T <: Any](
-      name: String,
-      clazz: Class[F],
-      bodyCode: String,
-      returnType: TypeInformation[T])
-    : GeneratedFunction[F, T] = {
-    val funcName = newName(name)
-
-    // Janino does not support generics, that's why we need
-    // manual casting here
-    val samHeader =
-      // FlatMapFunction
-      if (clazz == classOf[FlatMapFunction[_, _]]) {
-        val baseClass = classOf[RichFlatMapFunction[_, _]]
-        val inputTypeTerm = boxedTypeTermForTypeInfo(input1)
-        (baseClass,
-          s"void flatMap(Object _in1, org.apache.flink.util.Collector $collectorTerm)",
-          List(s"$inputTypeTerm $input1Term = ($inputTypeTerm) _in1;"))
-      }
-
-      // MapFunction
-      else if (clazz == classOf[MapFunction[_, _]]) {
-        val baseClass = classOf[RichMapFunction[_, _]]
-        val inputTypeTerm = boxedTypeTermForTypeInfo(input1)
-        (baseClass,
-          "Object map(Object _in1)",
-          List(s"$inputTypeTerm $input1Term = ($inputTypeTerm) _in1;"))
-      }
-
-      // FlatJoinFunction
-      else if (clazz == classOf[FlatJoinFunction[_, _, _]]) {
-        val baseClass = classOf[RichFlatJoinFunction[_, _, _]]
-        val inputTypeTerm1 = boxedTypeTermForTypeInfo(input1)
-        val inputTypeTerm2 = boxedTypeTermForTypeInfo(input2.getOrElse(
-          throw new CodeGenException("Input 2 for FlatJoinFunction should not be null")))
-        (baseClass,
-          s"void join(Object _in1, Object _in2, org.apache.flink.util.Collector $collectorTerm)",
-          List(s"$inputTypeTerm1 $input1Term = ($inputTypeTerm1) _in1;",
-               s"$inputTypeTerm2 $input2Term = ($inputTypeTerm2) _in2;"))
-      }
-
-      // ProcessFunction
-      else if (clazz == classOf[ProcessFunction[_, _]]) {
-        val baseClass = classOf[ProcessFunction[_, _]]
-        val inputTypeTerm = boxedTypeTermForTypeInfo(input1)
-        (baseClass,
-          s"void processElement(Object _in1, " +
-            s"org.apache.flink.streaming.api.functions.ProcessFunction.Context $contextTerm," +
-            s"org.apache.flink.util.Collector $collectorTerm)",
-          List(s"$inputTypeTerm $input1Term = ($inputTypeTerm) _in1;"))
-      }
-      else {
-        // TODO more functions
-        throw new CodeGenException("Unsupported Function.")
-      }
-
-    val funcCode = j"""
-      public class $funcName
-          extends ${samHeader._1.getCanonicalName} {
-
-        ${reuseMemberCode()}
-
-        public $funcName() throws Exception {
-          ${reuseInitCode()}
-        }
-
-        ${reuseConstructorCode(funcName)}
-
-        @Override
-        public void open(${classOf[Configuration].getCanonicalName} parameters) throws Exception {
-          ${reuseOpenCode()}
-        }
-
-        @Override
-        public ${samHeader._2} throws Exception {
-          ${samHeader._3.mkString("\n")}
-          ${reusePerRecordCode()}
-          ${reuseInputUnboxingCode()}
-          $bodyCode
-        }
-
-        @Override
-        public void close() throws Exception {
-          ${reuseCloseCode()}
-        }
-      }
-    """.stripMargin
-
-    GeneratedFunction(funcName, returnType, funcCode)
-  }
-
-  /**
-    * Generates a values input format that can be passed to Java compiler.
-    *
-    * @param name Class name of the input format. Must not be unique but has to be a
-    *             valid Java class identifier.
-    * @param records code for creating records
-    * @param returnType expected return type
-    * @tparam T Return type of the Flink Function.
-    * @return instance of GeneratedFunction
-    */
-  def generateValuesInputFormat[T <: Row](
-      name: String,
-      records: Seq[String],
-      returnType: TypeInformation[T])
-    : GeneratedInput[GenericInputFormat[T], T] = {
-    val funcName = newName(name)
-
-    addReusableOutRecord(returnType)
-
-    val funcCode = j"""
-      public class $funcName extends ${classOf[GenericInputFormat[_]].getCanonicalName} {
-
-        private int nextIdx = 0;
-
-        ${reuseMemberCode()}
-
-        public $funcName() throws Exception {
-          ${reuseInitCode()}
-        }
-
-        @Override
-        public boolean reachedEnd() throws java.io.IOException {
-          return nextIdx >= ${records.length};
-        }
-
-        @Override
-        public Object nextRecord(Object reuse) {
-          switch (nextIdx) {
-            ${records.zipWithIndex.map { case (r, i) =>
-              s"""
-                 |case $i:
-                 |  $r
-                 |break;
-               """.stripMargin
-            }.mkString("\n")}
-          }
-          nextIdx++;
-          return $outRecordTerm;
-        }
-      }
-    """.stripMargin
-
-    GeneratedInput(funcName, returnType, funcCode)
-  }
-
-  /**
-    * Generates a [[TableFunctionCollector]] that can be passed to Java compiler.
-    *
-    * @param name Class name of the table function collector. Must not be unique but has to be a
-    *             valid Java class identifier.
-    * @param bodyCode body code for the collector method
-    * @param collectedType The type information of the element collected by the collector
-    * @return instance of GeneratedCollector
-    */
-  def generateTableFunctionCollector(
-      name: String,
-      bodyCode: String,
-      collectedType: TypeInformation[Any])
-    : GeneratedCollector = {
-
-    val className = newName(name)
-    val input1TypeClass = boxedTypeTermForTypeInfo(input1)
-    val input2TypeClass = boxedTypeTermForTypeInfo(collectedType)
-
-    val funcCode = j"""
-      public class $className extends ${classOf[TableFunctionCollector[_]].getCanonicalName} {
-
-        ${reuseMemberCode()}
-
-        public $className() throws Exception {
-          ${reuseInitCode()}
-        }
-
-        @Override
-        public void collect(Object record) throws Exception {
-          super.collect(record);
-          $input1TypeClass $input1Term = ($input1TypeClass) getInput();
-          $input2TypeClass $input2Term = ($input2TypeClass) record;
-          ${reuseInputUnboxingCode()}
-          $bodyCode
-        }
-
-        @Override
-        public void close() {
-        }
-      }
-    """.stripMargin
-
-    GeneratedCollector(className, funcCode)
   }
 
   /**
