+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/codegen/ExpressionReducer.scala
@@ -33,7 +33,7 @@ import org.apache.flink.types.Row
 import scala.collection.JavaConverters._
 
 /**
-  * Evaluates constant expressions using Flink's [[CodeGenerator]].
+  * Evaluates constant expressions using Flink's [[FunctionCodeGenerator]].
   */
 class ExpressionReducer(config: TableConfig)
   extends RelOptPlanner.Executor with Compiler[MapFunction[Row, Row]] {
@@ -77,7 +77,7 @@ class ExpressionReducer(config: TableConfig)
     val resultType = new RowTypeInfo(literalTypes: _*)
 
     // generate MapFunction
-    val generator = new CodeGenerator(config, false, EMPTY_ROW_INFO)
+    val generator = new FunctionCodeGenerator(config, false, EMPTY_ROW_INFO)
 
     val result = generator.generateResultExpression(
       resultType,
